feat(user-api): :sparkles: add user registration and update endpoints with validation and logging
