+++ Apps/Accounts/api/user/serializers/user.py
@@ -1,0 +1,37 @@
+from rest_framework import serializers
+from Apps.Accounts.models import User
+
+
+class UserSerializer(serializers.ModelSerializer):
+    password = serializers.CharField(write_only=True, required=False, style={'input_type': 'password'})
+    confirm_password = serializers.CharField(write_only=True, required=False, style={'input_type': 'password'})
+
+    class Meta:
+        model = User
+        fields = ['username', 'email', 'password',
+                  'confirm_password', 'first_name', 'last_name',
+                  'phone', 'address', 'city', 'country', 'is_seller']
+
+    def validate(self, data):
+        if self.instance and not any([data.get('password'), data.get('confirm_password')]):
+            return data
+        password = data.get('password')
+        confirm_password = data.get('confirm_password')
+        if password != confirm_password:
+            raise serializers.ValidationError("Las contraseñas no coinciden.")
+        return data
+
+    def create(self, validated_data):
+        validated_data.pop('confirm_password', None)
+        return User.objects.create_user(**validated_data)
+
+    def update(self, instance, validated_data):
+        password = validated_data.pop('password', None)
+        validated_data.pop('confirm_password', None)
+        instance = super().update(instance, validated_data)
+
+        if password:
+            instance.set_password(password)
+            instance.save()
+
+        return instance
