+++ Apps/Accounts/api/user/urls.py
@@ -1,7 +1,10 @@
 from django.urls import path
 
-from Apps.Accounts.api import UserDeleteView
+from Apps.Accounts.api import UserDeleteView, RegisterUserView, UpdateUserView
 
 urlpatterns = [
     path('delete/', UserDeleteView.as_view(), name='delete_user'),
+    path('register/', RegisterUserView.as_view(), name='register'),
+    path('update/', UpdateUserView.as_view(), name='user_update'),
+
 ]
