--- Apps/Accounts/views/user/__init__.py
@@ -1,1 +1,0 @@
-from .user import *
